Add support for opbect operation checks.
With a strict separation between YAML and Rego, it can be
awkward to express checks that apply to the status of an
object update. If we allow checks to be directly attached to
the object YAML, then we can supply the result of the object
update as the input document to the check. This also allows us
to supply a default check that can be overridden for specific
objects. Although the syntax for this is awkward, flipping
checks per-object is important, and the brevity of having a
default is important too.