Move runner code into a step.
Outside of a step helper, we should not assume any state, because we
could have encountered a fatal error in the previous step. Since the
step helper omits the step if there is a fatal error pending, we have
to start a step and skip it from within if we didn't want to perform it.