Add support for importing Rego documents.
Add the `--policies` flag to load Rego policies into the runtime. This
required a lot of restructuring because we want to compile the Rego
files at the right time so that we can exit with syntax errors. The Rego
compiler doesn't let out add new compilation units, so we have to front
load everything as much as possible.