Improve result extraction.
Update result extraction to always include a message reporting the name
of the query that produced the result.

If a query produces an empty result set, we no longer populate a test
result (because that's not actually a failure). This fixes a regression
introduced on the last attempt to improve result extraction.

Adds support for results that are slices of strings, which allows Rego
rules to simplify multi-line response messages.